app/backend/requirements.txt
Flask-JWT-Extended
Flask-CORS
flask_apscheduler
flask-sqlalchemy
SQLAlchemy
bcrypt
ethiopian-date
python-dotenv
mysqlclient
mysql-connector-python
pytest
schedule
-U marshmallow
flask-marshmallow
pyethiodate
